refactoring and compiling app
