+++ src/App.css
@@ -7,99 +7,3 @@ body{
   width: 40%;
 }
 
-.Header{
-  display: flex;
-  flex-direction: column;
-}
-
-.pageSizeBox{
-  display: flex;
-  margin-bottom: 10px;
-}
-.pageSizeBox label{
-  margin-left: auto;
-  margin-right: 5px;
-}
-
-
-.todoBox h2{
-  text-align: center;
-}
-
-.todoBox input[type=text]{
-  align-self: center;
-  width: 40%;
-  padding:1%;
-  -webkit-border-radius: 5px;
-  -moz-border-radius: 5px;
-  border-radius: 5px;
-  margin-bottom:20px;
-}
-
-.warningMessage{
-  color:000;
-  font-weight: bold;
-  text-align: center;
-  background-color: salmon;
-  border: 2px solid red;
-}
-
-.todoBox input[type=text]:focus {
-  outline: none;
-}
-
-.todoBox ul, .todoBox li{
-  display: block;
-}
-.todoBox ul{
-  padding-left: 0px;
-}
-
-.todoBox li{
-  padding: 10px 0;
-  border-bottom: 2px solid #DDD;
-}
-
-.todoBox li input[type="checkbox"]{
-  margin-right: 5%;
-}
-
-.todoBox li span{
-  float: right;
-}
-
-.todoBox li img{
-  width: 30px;
-  height: 30px;
-  border-radius: 50%;
-  float: right;
-  margin-left: 20px;
-}
-
-
-
-.taskBottom{
-  display: flex;
-  margin-top:20px;
-}
-
-.taskBottom .RBox{
-  margin-left: auto;
-  color:#000;
-  text-decoration: underline;
-}
-
-
-.pagination_box{
-  margin-top:20px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.pagination_box button{
-  border-radius: 5px;
-  border:2px solid #333;
-  background-color: #EEE;
-  font-weight: bold;
-}
