+++ src/components/Header/Header.js
@@ -1,4 +1,6 @@
-const handleAddInputKeyPress = (e, tasks, setTaskToAdd) => {
+import './Header.scss';
+
+const handleAddInputKeyPress = (e, setTaskToAdd) => {
   if(e.key === 'Enter'){
     let newTask = {title: e.target.value, completed: false, who: "Me"};
     setTaskToAdd(newTask);
@@ -9,6 +11,7 @@ const handleAddInputKeyPress = (e, tasks, setTaskToAdd) => {
 let handleSearchKeyPress = (e, setSearch) => {
   if(e.code === 'Enter'){
     setSearch(e.target.value);
+    e.target.value = "";
   }
 }
 
@@ -18,10 +21,15 @@ const Header = ({tasks, pagination}) => {
         <h2>What do you want to do today?</h2>
 
         <div className="Header">
-          <input className="addInput" type="text" onKeyPress={(e) =>  handleAddInputKeyPress(e, tasks.fullTasks, tasks.setTaskToAdd) }  placeholder="Perss'Enter' to add task"></input>
+          <input className="addInput" type="text" onKeyPress={(e) =>  handleAddInputKeyPress(e, tasks.setTaskToAdd) }  placeholder="Perss'Enter' to add task"></input>
           <input className="searchInput" type="text" placeholder="Search" onKeyPress={(e) => handleSearchKeyPress(e, tasks.setSearch) } ></input>
         </div>
-
+        
+        {tasks.search !== '' && <div>
+          Searching '<b>{tasks.search}</b>' 
+          <button onClick={() => tasks.setSearch('') }>Clear</button>
+        </div> }
+        
         <div className="pageSizeBox">
           <label>Page Size</label>
           <select onChange={(e) => {pagination.setPageSize(e.target.value); pagination.setPage(0)}}>
