--- src/components/Pagination.js
@@ -1,7 +1,0 @@
-const Pagination = () => {
-    return (
-        <botton>Previous</botton>
-    )
-}
-
-export default Pagination;
