+++ src/components/TaskList/TaskOverWorkWarning/TaskOverWorkWarning.js
@@ -1,3 +1,5 @@
+import './TaskOverWorkWarning.scss';
+
 function TaskOverworkWarning({tasks}){
     let tasksLeft = tasks.filter(t => !t.completed).length;
     return (tasksLeft >= 5)? <div className="warningMessage">Hay 5 o mas tareas por hacer</div> : '';
