+++ src/components/TaskList/TaskOverWorkWarning/TaskOverWorkWarning.scss
@@ -1,0 +1,9 @@
+
+.warningMessage{
+    color:000;
+    font-weight: bold;
+    text-align: center;
+    background-color: salmon;
+    border: 2px solid red;
+  }
+
